requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
requests==2.31.0
python-dotenv==1.0.0
firebase-admin==6.2.0
docxtpl==0.20.2
google-genai>=1.0.0
trafilatura>=2.0.0
Pillow>=10.0.0
